Update tests and fix GetImages bug
Add additional directories and files for testing the GetImages class,
including files with the same names in different directories. Fix a bug
in GetImages that missed files with the same name in different
directories. GetImages now prepends the Job ID to each file name. Added
a method to remove the Job ID to clean up after tests.